test: extract shared scaffolding into test/helpers.ts
FakeExtensionApi was copy-pasted verbatim across four suites, and the
TestContext types plus context() factory across three-and-a-half (the
temp-failure copy was a strict subset). rmRetry, makeRepository and
privateRefs were duplicated on top of that. Hoist them into one
test/helpers.ts: -409/+24 in the suites for +133 shared lines.

Notes on the non-mechanical bits:

- lifecycle keeps its longer teardown as a one-line wrapper over the
  shared rmRetry(path, attempts = 6), so its 40 callsites and the
  10-attempt behaviour are unchanged.
- checkpoints-temp-failures imports git as rawGit to leave its callsites
  alone, and drops its narrower TestContext: the shared context() leaves
  branch/entries empty, so getBranch() -> [] and getEntry() -> undefined
  still hold.
- Repo temp prefixes collapse to one omp-undo-redo-repo-, which only
  costs the per-suite label on leftover temp directories.
- Shared TestContext declares cwd, which every copy assigned without
  declaring.
- rmRetry sleeps via node:timers/promises rather than wrapping
  setTimeout in a Promise executor.

checkpoint-owners keeps its own makeRepository: it returns a runner and
common dir and asserts exit codes, so sharing it would need an options
bag and a union return.

Full suite passes (131 passed, 2 skipped); lint, prettier and typecheck
clean. test/ is excluded from tsconfig, but an ad-hoc tsc over these
files drops from 9 pre-existing errors to 4, with none introduced.
