+++ test/bounded-capture.test.ts
@@ -5,8 +5,7 @@ import { afterAll, describe, expect, it } from "vitest";
 import { createEnvGitRunner, createGitRunner } from "../src/core/git-runner.js";
 import type { GitRunner } from "../src/core/types.js";
 import ompUndoRedo, { type OmpUndoRedoDependencies } from "../src/index.js";
-
-type Handler = (...args: unknown[]) => unknown;
+import { context, FakeExtensionApi, rmRetry } from "./helpers.js";
 
 const testStoreRoot = join(tmpdir(), `omp-undo-redo-test-store-${process.pid}`);
 process.env.OMP_UNDO_REDO_STORE_DIR = testStoreRoot;
@@ -14,99 +13,6 @@ process.env.OMP_UNDO_REDO_STORE_DIR = testStoreRoot;
 afterAll(async () => {
   await rm(testStoreRoot, { recursive: true, force: true });
 });
-
-async function rmRetry(path: string, attempts = 6): Promise<void> {
-  for (let attempt = 0; attempt < attempts; attempt += 1) {
-    try {
-      await rm(path, { recursive: true, force: true });
-      return;
-    } catch (error) {
-      if (attempt === attempts - 1) throw error;
-      await new Promise((resolve) => setTimeout(resolve, 200));
-    }
-  }
-}
-
-type TestEntry = {
-  id: string;
-  parentId: string | null;
-  type: string;
-  message?: { role?: string };
-  customType?: string;
-};
-
-type TestContext = {
-  branch: TestEntry[];
-  entries: TestEntry[];
-  leaf: string;
-  sessionManager: {
-    getSessionId(): string;
-    getLeafId(): string;
-    getBranch(): TestEntry[];
-    getEntry(id: string): TestEntry | undefined;
-  };
-  navigateTree(targetId: string): Promise<{ cancelled: boolean }>;
-  waitForIdle(): Promise<void>;
-  isIdle(): boolean;
-  ui: {
-    notifications: Array<{ message: string; level: string }>;
-    notify(message: string, level: string): void;
-  };
-};
-
-class FakeExtensionApi {
-  private readonly handlers = new Map<string, Handler>();
-  private readonly commands = new Map<string, Handler>();
-
-  on(event: string, handler: Handler): void {
-    this.handlers.set(event, handler);
-  }
-
-  registerCommand(name: string, config: { handler: Handler }): void {
-    this.commands.set(name, config.handler);
-  }
-
-  async runCommand(name: string, context: TestContext): Promise<void> {
-    const handler = this.commands.get(name);
-    if (!handler) throw new Error(`No command registered for ${name}`);
-    await handler("", context);
-  }
-
-  async emit(
-    event: string,
-    context: TestContext,
-    payload?: Record<string, unknown>,
-  ): Promise<void> {
-    const handler = this.handlers.get(event);
-    if (!handler) throw new Error(`No handler registered for ${event}`);
-    await handler(payload ?? { type: event }, context);
-  }
-}
-
-function context(cwd: string, sessionId: string): TestContext {
-  const value: TestContext = {
-    cwd,
-    leaf: "leaf",
-    branch: [],
-    entries: [],
-    sessionManager: {
-      getSessionId: () => sessionId,
-      getLeafId: () => value.leaf,
-      getBranch: () => value.branch,
-      getEntry: (id) => value.entries.find((entry) => entry.id === id),
-    },
-    navigateTree: async () => ({ cancelled: true }),
-    waitForIdle: async () => {},
-    isIdle: () => true,
-    ui: {
-      notifications: [],
-      notify(message, level) {
-        value.ui.notifications.push({ message, level });
-      },
-    },
-  };
-  return value;
-}
 
 /** A runner factory whose git-add invocations take `delayMs` to complete.
  *  The returned waiter lets tests await an exact number of completed adds,
